+++ mypyc/ops_int.py
@@ -1,22 +1,5 @@
-from typing import Dict, List, Callable
-
-from mypyc.ops import (
-    OpDescription, PrimitiveOp2, int_rprimitive, RType, EmitterInterface, ERR_NEVER
-)
-
-
-binary_ops = {}  # type: Dict[str, List[OpDescription]]
-
-
-def binary_op(op: str,
-              arg_types: List[RType],
-              result_type: RType,
-              error_kind: int,
-              format_str: str,
-              emit: Callable[[EmitterInterface, PrimitiveOp2], None]) -> None:
-    ops = binary_ops.setdefault(op, [])
-    desc = OpDescription(op, arg_types, result_type, error_kind, format_str, emit)
-    ops.append(desc)
+from mypyc.ops import PrimitiveOp2, int_rprimitive, RType, EmitterInterface, ERR_NEVER
+from mypyc.ops_primitive import binary_op
 
 
 def int_binary_op(op: str, c_func_name: str) -> None:
